+++ src/Functionality/Analysis/binaryValidation.py
