+++ src/Functionality/Analysis/staticAnalysis.py
